# Switch to raw SQL from WP_Date_Query